// testui/src/app/models/sumarry-reconciliation-response.ts
export interface SummaryReconciliationResponse {
  reconciliationDate: String;
  reconciliationResult: String;
  responseHeader: String;
  totalPostAuthAmount: String;
  totalPostAuthCount: String;
  totalPostAuthReverseAmount: String;
  totalPostAuthReverseCount: String;
  totalPreAuthAmount: String;
  totalPreAuthCount: String;
  totalPreAuthReverseAmount: String;
  totalPreAuthReverseCount: String;
  totalRefundAmount: String;
  totalRefundCount: String;
  totalReverseAmount: String;
  totalReverseCount: String;
  totalSaleAmount: String;
  totalSaleCount: String;
}